pkg/otlp/metrics: add remapping code
This change adds the processing code for extracting Datadog metrics from
container.* and system.* metrics with the intention of centralising this
logic to be re-used in the backend intake.

The operation is restricted by the newly added WithRemapping translator
configuration option, so that it does not affect the translator's usage as part
of the Datadog Agent, where no remapping is necessary because the agent
creates the necessary metrics.