[OTEL-1325] Ignore diff for values with duplicate timestamps
When we receive two subsequent values with the same timestamp, the second value received will be ignored for diffs. Currently we still compute the diff.

This PR also fixes an issue where a point could be considered and computed as the first value, if it's timestamp was older than the timestamp of the point we currently have in the cache. Now, the first point calculation will not be considered for points which are older or equal to the current point in cache (having a current point in cache indicates that it is not the first point).

This is the First PR to compute rates. In the case of rates, having the same timestamp would mean that `dt` == 0 so we cannot compute the rate.