[OTEL-1342][pkg/otlp/metrics] Add kafka metrics mappings (#239)
* [OTEL-1325][pkg/otlp/metrics] Add support for computing rates

This PR adds support for rate as gauge metrics. For any monotonic metric present in the rateAsGauge map, we will compute the rate and submit the metric as a gauge. First val for rateAsGauge metrics are cached but not submitted as first val.

This PR is necessary to add support for kafka metrics that compute a rate and are submitted as a gauge.

* make fmt

* remove unnecessary conversion

* [OTEL-1342][pkg/otlp/metrics] Add kafka metrics mappings

* use kafka metric in metric translator test

* Add more test cases

* refactor tests and add jvm

* add test case for dynamic attr missing

* changelog + licenses

* licenses

* add licenses

* add collector to depguard

* address feedback